support/set test for csv import customer
