--- app/services/ApplicationService.rb
@@ -1,5 +1,0 @@
-class ApplicationService
-	def self.call(*args, &block)
-		new(*args, &block).call
-	end
-end
