--- app/services/ImportCustomerCSV.rb
@@ -1,15 +1,0 @@
-class ImportCustomerCSV
-	include CSVImporter
-
-	model Customer
-
-	# Maps the csv columns to attributes
-	column :cellphone
-	column :cpf
-	column :email
-	column :name
-	column :phone
-	column :account_id  
-
-	identifier :email 
-end
